continue reordering code
